Made projects page mobile friendly, or at least starting towards that area.
Mobile friendly will be very helpful for when trying to show to other people on the go, and it's responsive so i can send it anyone and it should display properly. One major thing was media queries in which I could change the projects page and home page to be using a column approach instead of a row approach.